Preset implemented change message processing

+++ app/src/main/java/cz/fjerabek/thr10controller/data/Property.kt
@@ -1,5 +1,7 @@
 package cz.fjerabek.thr10controller.data
 
+import kotlin.reflect.KClass
+
 @Target(AnnotationTarget.PROPERTY)
-annotation class Property(val propertyId : Int) {
+annotation class Property(val propertyId : Byte) {
 }
